fix: move usage of locatstorage into pokemonsearchpage component
