+++ src/PokemonSearchPage/PokemonSearchPage.test.tsx
@@ -21,3 +21,15 @@ test('updating searchString and pass it into TableView', async () => {
   const char = await screen.findByText('charmander');
   expect(char).toBeInTheDocument();
 });
+
+test('save searchString into localStorage on Search click', async () => {
+  render(<PokemonSearchPage />);
+  const input = screen.getByRole('textbox');
+  const searchButton = screen.getByRole('button', { name: /search/i });
+
+  await userEvent.clear(input);
+  await userEvent.type(input, 'pikachu');
+  await userEvent.click(searchButton);
+
+  expect(localStorage.getItem('searchInput')).toBe('pikachu');
+});
