+++ src/PokemonSearchPage/PokemonSearchPage.tsx
@@ -1,21 +1,23 @@
 import { useState } from 'react';
-import Search from '../Search/Search';
-import TableView from '../TableView/TableView';
-import ErrorBoundary from '../ErrorBoundary/ErrorBoundary';
+import Search from '../components/Search/Search';
+import TableView from '../components/TableView/TableView';
+import ErrorBoundary from '../components/ErrorBoundary/ErrorBoundary';
 
 export default function PokemonSearchPage() {
-  //todo: specifi searchstr only in there or only in Search component
   const [searchData, setSearchData] = useState(
     localStorage.getItem('searchInput') || ''
   );
 
-  const handleSearchData = (data: string) => setSearchData(data);
+  const handleSearchData = (data: string) => {
+    setSearchData(data);
+    localStorage.setItem('searchInput', data);
+  };
 
   return (
     <>
       <header>
         <ErrorBoundary>
-          <Search sendSearchUp={handleSearchData} />
+          <Search searchStr={searchData} sendSearchUp={handleSearchData} />
         </ErrorBoundary>
       </header>
       <main>
